Documentation :) + Adding TaxonomyTree (kind of wrapper to Node class)
Added documentation.
Added NodeID to the Node class.
Added TaxonomyTree as a wrapper over the node, so that it can keep track
of the single tree created for the application, using its RootNode and
the count of the nodes in the tree.

Next Step: Using PrefixMap class to create the Prefix Map for the
TaxonomyTree.